pyproject.toml
[tool.poetry]
name = "converterformats"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
dicttoxml = "^1.7.4"
toml = "^0.10.2"
pyyaml = "*"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
